requirements.txt
python-dotenv
pydantic_settings
fastapi
uvicorn
passlib
Secweb